pyproject.toml
[tool.poetry]
name = "dashboard_env"
version = "0.1.0"
description = ""
authors = ["Guinslym <guinslym@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "2.2.19"
python-dateutil = "^2.8.1"
python-dotenv = "^0.17.1"
BeautifulSoup4 = "^4.9.3"
lh3api = "^0.2.0"
pandas = "1.1.5"
XlsxWriter = "^1.4.3"
django-extensions = "^3.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^4.4.0"
pytest-watch = "^4.2.0"
pytest-sugar = "^0.9.4"
pytest-cov = "^2.12.1"
django-sphinxdoc = "^2.0.0"
sphinx = "^4.0.2"
ipython = "^7.24.1"
black = "^21.6b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
